+++ packages/numbers/src/3-candidate/index.ts
@@ -1,5 +1,6 @@
 export { generateRandom } from './generateRandom.ts';
 export { isIntegerString, safeParseInteger } from './integer-string.ts';
+export { isNumericString, safeParseNumber } from './numeric-string.ts';
 export { pickInteger } from './pickInteger.ts';
 export { round } from './round.ts';
 export { scale } from './scale.ts';